Revision One: June 4, 2013
As a note to Mateusz:

GitHub offers branches. I have made one for you and for me so that we
can work on separate versions of the code without harming the original.
Once we have stable work on either side or both sides containing
features that work with each other, they will be published to the master
branch.

Always publish at least sudo-stable commits to master...everything else
that you want to keep in your personal endeavors, leave to your branch
until you are sure it works and you want to merge.